+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Disable per-player tech owing to on-going issues pending the next game patch
 - GWO panel displays your co-op settings when you call for reinforcements
+- Support multiple human players on the GWO panel
 
 ## v5.87.1 - 2026-06-26